+++ .ci/test.sh
@@ -3,5 +3,5 @@ set -e -v
 
 if [[ $1 == "script" ]]; then
   pytest -n 2 --pyargs nengo
-  pytest -n 2 --durations 20 nengo_dl
+  pytest -n 2 --durations 20 $TEST_ARGS nengo_dl
 fi